requirements.txt
Django==3.2.6
black==21.8b0
factory-boy==3.2.0